# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CounselorListResponse:
    properties:
      counselors:
        items:
          $ref: '#/definitions/dto.CounselorResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.CounselorRequest:
    properties:
      gender:
        type: string
      institution_code:
        type: string
      language_skills:
        type: string
      license_number:
        type: string
      name:
        type: string
      psychology_school:
        type: string
      specialties:
        type: string
      treatment_methods:
        type: string
      work_location:
        type: string
      work_unit:
        type: string
    required:
    - license_number
    - name
    type: object
  dto.CounselorResponse:
    properties:
      created_at:
        type: string
      gender:
        type: string
      id:
        type: string
      institution_code:
        type: string
      language_skills:
        type: string
      license_number:
        type: string
      name:
        type: string
      psychology_school:
        type: string
      specialties:
        type: string
      treatment_methods:
        type: string
      updated_at:
        type: string
      work_location:
        type: string
      work_unit:
        type: string
    type: object
...
  vo.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      error:
        type: string
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  vo.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: mindhelp.onrender.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MindHelp 心理健康支援應用程式後端 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MindHelp Backend API
  version: "1.0"
paths:
...
  /admin/counselors:
    post:
      consumes:
      - application/json
      description: 創建新的諮商師記錄
      parameters:
      - description: 諮商師資訊
        in: body
        name: counselor
        required: true
        schema:
          $ref: '#/definitions/dto.CounselorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CounselorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 創建諮商師
      tags:
      - counselors
  /admin/counselors/{id}:
    delete:
      consumes:
      - application/json
      description: 刪除諮商師記錄
      parameters:
      - description: 諮商師ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 刪除諮商師
      tags:
      - counselors
    put:
      consumes:
      - application/json
      description: 更新諮商師資訊
      parameters:
      - description: 諮商師ID
        in: path
        name: id
        required: true
        type: string
      - description: 諮商師資訊
        in: body
        name: counselor
        required: true
        schema:
          $ref: '#/definitions/dto.CounselorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CounselorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 更新諮商師
      tags:
      - counselors
...
  /counselors:
    get:
      consumes:
      - application/json
      description: 獲取諮商師列表，支援分頁和搜索
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: page_size
        type: integer
      - description: 搜索關鍵字
        in: query
        name: search
        type: string
      - description: 工作地點
        in: query
        name: work_location
        type: string
      - description: 專業領域
        in: query
        name: specialty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CounselorListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 獲取諮商師列表
      tags:
      - counselors
  /counselors/{id}:
    get:
      consumes:
      - application/json
      description: 根據ID獲取諮商師詳細資訊
      parameters:
      - description: 諮商師ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CounselorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 獲取單個諮商師
      tags:
      - counselors
...
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer " 加上 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"